+++ simulator/src/com/verictas/pos/simulator/Object.java
@@ -4,8 +4,8 @@ import java.lang.*;
 
 public class Object {
     public double mass;
-    public Vector3f position;
-    public Vector3f speed;
+    public Vector3d position;
+    public Vector3d speed;
     private double gravitationalConstant = 6.67384E-11;
 
     /**
@@ -14,7 +14,7 @@ public class Object {
      * @param position The position vector of the object
      * @param speed The speed vector of the object
      */
-    public Object(double mass, Vector3f position, Vector3f speed) {
+    public Object(double mass, Vector3d position, Vector3d speed) {
         this.mass = mass;
         this.position = position;
         this.speed = speed;
@@ -24,7 +24,7 @@ public class Object {
      * Sets the speed vector of an object
      * @param speed Current speed vector
      */
-    public void setSpeed(Vector3f speed) {
+    public void setSpeed(Vector3d speed) {
         this.speed = speed;
     }
 
@@ -32,7 +32,7 @@ public class Object {
      * Sets the position vector of an object.
      * @param position Current position vector
      */
-    public void setPosition(Vector3f position) {
+    public void setPosition(Vector3d position) {
         this.position = position;
     }
 
@@ -47,22 +47,22 @@ public class Object {
     /**
      * Calculates the force of the passed object on the current object.
      * @param secondObject The passed object
-     * @return Vector3f The gravitational force
+     * @return Vector3d The gravitational force
      */
-    public Vector3f getForceOnObject(Object secondObject) {
+    public Vector3d getForceOnObject(Object secondObject) {
         double scale = gravitationalConstant * ((this.mass * secondObject.mass) / Math.pow(getDistance(secondObject).length(), 3.0));
-        Vector3f force = getDistance(secondObject);
-        force.scale((float) scale);
+        Vector3d force = getDistance(secondObject);
+        force.scale(scale);
         return force;
     }
 
     /**
      * Get the vector distance between the current position vector and the position vector of the passed object.
      * @param secondObject The passed object.
-     * @return Vector3f The distance vector
+     * @return Vector3d The distance vector
      */
-    private Vector3f getDistance(Object secondObject) {
-        Vector3f distance = new Vector3f(0,0,0); // Empty
+    private Vector3d getDistance(Object secondObject) {
+        Vector3d distance = new Vector3d(0,0,0); // Empty
         distance.sub(this.position, secondObject.position);
         return distance;
     }
